+++ packages/dds-components/src/components/OverflowMenu/OverflowMenuGroup.tsx
@@ -1,7 +1,9 @@
 import {
+  type Dispatch,
   Children as ReactChildren,
   type ReactElement,
   type ReactNode,
+  type SetStateAction,
   cloneElement,
   isValidElement,
   useId,
@@ -9,45 +11,93 @@ import {
   useState,
 } from 'react';
 
+import { OverflowMenuContextProvider } from './OverflowMenu.context';
+import {
+  type UseFloatPositionOptions,
+  useCombinedRef,
+  useFloatPosition,
+  useOnClickOutside,
+  useOnKeyDown,
+} from '../../hooks';
+
 export interface OverflowMenuGroupProps {
-  /**Ekstra logikk som kjøres når `<OverflowMenu />` åpnes og lukkes. */
-  onToggle?: () => void;
-  /**Ekstra logikk som kjøres når `<OverflowMenu />` åpnes. */
+  /**Om `<OverflowMenu>` er åpen ved første render.
+   * @default false
+   */
+  isInitiallyOpen?: boolean;
+  /**Implementerer kontrollert tilstand: forteller `<OverflowMenu>` om den skal være åpen. */
+  isOpen?: boolean;
+  /**Implementerer kontrollert tilstand: funksjon for å kontrollere `isOpen`. */
+  setIsOpen?: Dispatch<SetStateAction<boolean>>;
+  /**Callback når `<OverflowMenu>` åpnes. */
   onOpen?: () => void;
-  /**Ekstra logikk som kjøres når `<OverflowMenu />` lukkes. */
+  /**Callback når `<OverflowMenu>` lukkes. */
   onClose?: () => void;
-  /**Barn, anchor-elementet som første og `<OverflowMenu />` som andre. */
+  /**Barn, anchor-elementet som første og `<OverflowMenu>` som andre. */
   children: ReactNode;
-  /**Custom id for `<OverflowMenu />`. */
+  /**Custom id for `<OverflowMenu>`. */
   overflowMenuId?: string;
 }
 
 export const OverflowMenuGroup = ({
   children,
   onClose,
-  onToggle,
   onOpen,
+  setIsOpen: propSetIsOpen,
+  isOpen: propIsOpen,
+  isInitiallyOpen = false,
   overflowMenuId,
 }: OverflowMenuGroupProps) => {
-  const [isOpen, setIsOpen] = useState(false);
+  const [internalIsOpen, internalSetIsOpen] = useState(isInitiallyOpen);
+
+  const [isOpen, setIsOpen] = [
+    propIsOpen ?? internalIsOpen,
+    propSetIsOpen ?? internalSetIsOpen,
+  ];
   const toggle = () => setIsOpen(!isOpen);
   const close = () => setIsOpen(false);
-  const buttonRef = useRef<HTMLButtonElement>(null);
 
   const generatedId = useId();
   const uniqueOverflowMenuId = overflowMenuId ?? `${generatedId}-overflowMenu`;
 
+  const [floatOptions, setFloatOptions] = useState<UseFloatPositionOptions>();
+  const { refs, styles: positionStyles } = useFloatPosition(null, floatOptions);
+
+  const buttonRef = useRef<HTMLButtonElement>(null);
+  const anchorRef = refs.setReference;
+  const combinedAnchorRef = useCombinedRef(buttonRef, anchorRef);
+  const menuRef = useRef<HTMLDivElement>(null);
+  const floatingRef = refs.setFloating;
+  const combinedMenuRef = useCombinedRef(menuRef, floatingRef);
+
   const handleClose = () => {
-    onClose && onClose();
-    close();
+    if (isOpen) {
+      onClose && onClose();
+      close();
+    }
   };
 
   const handleToggle = () => {
-    onOpen && !isOpen && onOpen();
-    onClose && isOpen && onClose();
-    onToggle && onToggle();
+    !isOpen && onOpen?.();
+    isOpen && onClose?.();
     toggle();
   };
+
+  useOnClickOutside([menuRef.current, buttonRef.current], () => {
+    handleClose();
+  });
+
+  useOnKeyDown(['Esc', 'Escape'], () => {
+    if (isOpen) {
+      onClose && onClose();
+      close();
+      buttonRef.current?.focus();
+    }
+  });
+
+  useOnKeyDown(['Tab'], () => {
+    handleClose();
+  });
 
   const Children = ReactChildren.map(children, (child, childIndex) => {
     return (
@@ -58,19 +108,24 @@ export const OverflowMenuGroup = ({
             'aria-controls': uniqueOverflowMenuId,
             'aria-expanded': isOpen,
             onClick: handleToggle,
-            ref: buttonRef,
+            ref: combinedAnchorRef,
           })
-        : cloneElement(child as ReactElement, {
-            isOpen: isOpen,
-            id: uniqueOverflowMenuId,
-            onClose: handleClose,
-            onToggle: handleToggle,
-            anchorRef: buttonRef,
-          }))
+        : child)
     );
   });
 
-  return <>{Children}</>;
+  return (
+    <OverflowMenuContextProvider
+      isOpen={isOpen}
+      onClose={handleClose}
+      menuRef={combinedMenuRef}
+      setFloatOptions={setFloatOptions}
+      floatStyling={positionStyles.floating}
+      menuId={uniqueOverflowMenuId}
+    >
+      {Children}
+    </OverflowMenuContextProvider>
+  );
 };
 
 OverflowMenuGroup.displayName = 'OverflowMenuGroup';