+++ packages/dds-components/src/components/Tabs/Tabs.stories.tsx
@@ -14,7 +14,14 @@ export default {
   argTypes: {
     width: { control: 'text' },
     htmlProps: { control: false },
+    addTabButtonProps: { control: false },
     activeTab: { control: 'number' },
+  },
+  parameters: {
+    docs: {
+      story: { inline: true },
+      canvas: { sourceState: 'shown' },
+    },
   },
 } satisfies Meta<typeof Tabs>;
 
@@ -102,6 +109,63 @@ export const ActiveTab: Story = {
           <TabPanel>Innhold 3</TabPanel>
         </TabPanels>
       </Tabs>
+    );
+  },
+};
+
+export const WithAddTabButton: Story = {
+  render: args => {
+    const [tooManyText, setTooManyText] = useState('');
+    const [activeTab, setActiveTab] = useState(0);
+    const [tabsState, setTabsState] = useState<
+      Array<{
+        name: string;
+        content: string;
+      }>
+    >([
+      { name: ' Tab 1', content: 'Innhold 1' },
+      { name: ' Tab 2', content: 'Innhold 2' },
+    ]);
+
+    const addTab = () => {
+      if (tabsState.length > 10) {
+        setTooManyText('For mange faner 🥵');
+      } else {
+        const index = tabsState.length;
+        const tabNumber = index + 1;
+        setTabsState([
+          ...tabsState,
+          { name: `Tab ${tabNumber}`, content: `Innhold ${tabNumber}` },
+        ]);
+        setActiveTab(index);
+      }
+    };
+
+    return (
+      <>
+        {tooManyText}
+        <Tabs
+          {...args}
+          activeTab={activeTab}
+          onChange={tab => setActiveTab(tab)}
+          addTabButtonProps={{
+            children: 'Legg til',
+            'aria-label': 'Legg til ny fane',
+            onClick: addTab,
+          }}
+        >
+          <TabList>
+            {tabsState.map(item => (
+              <Tab key={item.name}>{item.name}</Tab>
+            ))}
+          </TabList>
+          <TabPanels>
+            {tabsState.map(item => (
+              <TabPanel key={item.name}>{item.content}</TabPanel>
+            ))}
+          </TabPanels>
+        </Tabs>
+      </>
     );
   },
 };