+++ packages/dds-components/src/components/Tabs/index.tsx
@@ -1,3 +1,4 @@
+export * from './AddTabButton';
 export * from './Tabs';
 export * from './Tab';
 export * from './TabList';