+++ packages/dds-tokens/dds/index.ts
@@ -1,2 +1,0 @@
-export * as ddsTokensCore from './build/js/Core/ddsTokens';
-export * as ddsTokensPublic from './build/js/Public/ddsTokens';