+++ packages/dds-tokens/src/index.ts
@@ -1,4 +1,4 @@
-import { ddsTokensCore, ddsTokensPublic } from '../dds';
+import { ddsTokensCore, ddsTokensPublic } from '../generated-tokens';
 
 export const ddsTokens = {
   core: { ...ddsTokensCore },