+++ packages/dds-tokens/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @norges-domstoler/dds-design-tokens
 
+## 7.3.0
+
+### Minor Changes
+
+- ef5e73e: Legger til token for semi-bold font weight.
+
+### Patch Changes
+
+- 37be771: Justert typografistiler for public theme heading size medium-xxsmall.
+- 9997ed1: Fjerner ubrukte font size tokens.
+
+## 7.2.0
+
+### Minor Changes
+
+- bc8eb1a: Ny tokentype og nye tokens: z-index. Brukes på elementer som legger seg over innhold. Med tokens definerer vi hierarki for når slike elementer dukker opp samtidig. Se oversikten i [Elsa Storybook](https://domstolene.github.io/designsystem/?path=/story/dds-design-tokens-tokens--z-index).
+
 ## 7.1.0
 
 ### Minor Changes