+++ packages/storybook-components/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @norges-domstoler/storybook-components
 
+## 0.0.15
+
+### Patch Changes
+
+- Updated dependencies [ef5e73e]
+- Updated dependencies [37be771]
+- Updated dependencies [9997ed1]
+  - @norges-domstoler/dds-design-tokens@7.3.0
+
+## 0.0.14
+
+### Patch Changes
+
+- Updated dependencies [bc8eb1a]
+  - @norges-domstoler/dds-design-tokens@7.2.0
+
 ## 0.0.13
 
 ### Patch Changes